[keycloak] Add the global.imageRegistry option
This change allows one to specify the image registry which will be used for each image individually, as well as a global image registry to override all image registries.

I used Bitnami's common chart to keep things standard and generic.

This change is extremely beneficial to those who use an airgapped environment with a private image registry.

Signed-off-by: Yarden Shoham <hrsi88@gmail.com>