Allow setting loadBalancerSourceRanges on the svc
When creating Service with type of LoadBalancer it results with Security
Group (or equivalent) open to 0.0.0.0/0. By allowing to define
`loadBalancerSourceRanges` one can limit that to a list of allowed
CIDRs. This is a must when a cloud policy of an organisation doesn't
allow opening security groups to 0.0.0.0/0.

Signed-off-by: Marcin Jedyk <m.xyzxyz@xyzxyzyxz.com>

Update charts/keycloak/Chart.yaml

Co-authored-by: Reinhard Nägele <unguiculus@gmail.com>