[keycloak] Avoid issues with older K8s versions
enableServiceLinks was only introduced with K8s v1.13, which means the
chart is incompatible with distributions that are still based on older
versions (like OpenShift 3).

This commit changes the template to only render `enableServiceLinks` if
it is enabled, allowing it the operator to disable it and restore
pre-1.13 compatibility.

Fixes #358

Signed-off-by: Manuel Hutter <manuel@vshn.ch>