Use with function to wrap imagePullSecrets in serviceaccount.yaml
When using an operator to inject _imagePullSecrets_ into deployed
Kubernetes resources, instead of configuring them with values
and running a `helm upgrade ...` later on, a conflict arises.

The chart produces `imagePullSecrets: []` when none are set. To avoid
this, simply wrap in a `with` clause like it is already done in
deployment.yaml and statefulset.yaml.

Signed-off-by: Sascha Peilicke <sascha@peilicke.de>